Menu changes

+++ VoxelZombieClient/Assets/IChunk.cs
@@ -2,7 +2,7 @@ using System;
 
 public interface IChunk
 {
-    byte this[int x, int y, int z] { get; set; }
+    UInt64 this[int x, int y, int z] { get; set; }
 
     bool dirty { get; set; }
 }
